backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-configurations==2.1
django-cors-headers==2.2.1
djangorestframework==3.8.2
djoser==1.1.5
psycopg2-binary==2.8.6
pytz==2018.4
django-celery-beat==1.5
celery==4.2.1
celery_once==3.0.1
whitenoise==3.3.1
django-celery-results==1.0.1
django-redis==4.7.0
requests==2.13.0
pyamqp
django-extensions
smk-python-sdk
betfairlightweight